Fix Visual Studio compiler warnings for x64 build
Address 200+ MSVC warnings across the codebase:

- Replace non-standard '\e' escape with '\033' for ANSI codes (comm.c)
- Add explicit (int) casts for double-to-int conversions in damage calculations
- Add explicit (int) casts for size_t-to-int conversions (strlen results)
- Add explicit (int) casts for pointer arithmetic (__int64 to int)
- Fix function pointer dereferencing syntax (special.c, olc_act.c)
- Migrate deprecated gethostbyaddr to getnameinfo (compat.c)
- Use _close() on Windows instead of deprecated close() (comm.c)
- Add (int) casts for SOCKET to int conversions on Windows x64
- Fix time_t to long conversions (board.c, db.c)

+++ src/classes/demon.c
@@ -899,8 +899,8 @@ void do_cone( CHAR_DATA *ch, char *argument ) {
 
     if ( ( sn = skill_lookup( "cone" ) ) < 0 ) return;
     spelltype = skill_table[sn].target;
-    level = ch->spl[spelltype] * 1.0;
-	level = level * 1.0;
+    level = (int)(ch->spl[spelltype] * 1.0);
+	level = (int)(level * 1.0);
 	act("You Blast $N with a cone of fire.",ch,NULL,victim,TO_CHAR);
 	act("$n Blasts you with a cone of fire.",ch,NULL,victim,TO_VICT);
     (*skill_table[sn].spell_fun) ( sn, level, ch, victim );
